Move GetMethodAndUrl func out to MethodAndUrlUtils unit
