+++ DevProxy.Plugins/Reporting/GraphMinimalPermissionsPlugin.cs
@@ -70,7 +70,7 @@ public sealed class GraphMinimalPermissionsPlugin(
             }
 
             var methodAndUrlString = request.Message;
-            var methodAndUrl = GraphUtils.GetMethodAndUrl(methodAndUrlString);
+            var methodAndUrl = MethodAndUrlUtils.GetMethodAndUrl(methodAndUrlString);
             if (methodAndUrl.Method.Equals("OPTIONS", StringComparison.OrdinalIgnoreCase))
             {
                 continue;
